+++ proofs/src/plonk/prover.rs
@@ -228,6 +228,7 @@ where
                         &pk.fixed_values,
                         &instance.instance_values,
                         &challenges,
+                        &mut rng,
                         transcript,
                     )
                 })